requirements.txt
# AICODE-NOTE: Минимальные зависимости для API и работы с БД
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
pydantic>=2,<3
openai
httpx
python-dotenv
pyyaml

